Address Matt's comments part 1
