+++ packages/optimizely-sdk/lib/core/project_config/entities.ts
@@ -33,3 +33,32 @@ export interface FeatureFlag {
 export interface FeatureKeyMap {
   [key: string]: FeatureFlag
 }
+
+export interface Variation {
+  id: string;
+  key: string;
+  featureEnabled: boolean;
+  variables: VariationVariable[];
+}
+
+export interface VariationVariable {
+  id: string;
+  value: string;
+}
+
+export interface Experiment {
+  id: string;
+  key: string;
+  status: string;
+  layerId: string;
+  variations: Variation[];
+  trafficAllocation: Array<{
+    entityId: string;
+    endOfRange: number;
+  }>;
+  audienceIds: string[];
+  // TODO[OASIS-6649]: Don't use object type
+  // eslint-disable-next-line  @typescript-eslint/ban-types
+  forcedVariations: object;
+  variationKeyMap?: {[key: string]: Variation}
+}
