+++ packages/optimizely-sdk/lib/index.d.ts
@@ -18,7 +18,7 @@ declare module '@optimizely/optimizely-sdk' {
   import { LogHandler, ErrorHandler } from '@optimizely/js-sdk-logging';
   import * as enums from '@optimizely/optimizely-sdk/lib/utils/enums';
   import * as logging from '@optimizely/optimizely-sdk/lib/plugins/logger';
-  import { EventDispatcher } from '@optimizely/js-sdk-event-processor';
+
   export { enums, logging };
 
   export function setLogger(logger: LogHandler | null): void;
@@ -31,12 +31,19 @@ declare module '@optimizely/optimizely-sdk' {
 
   export const eventDispatcher: EventDispatcher;
 
+  export interface DatafileOptions {
+    autoUpdate?: boolean;
+    updateInterval?: number;
+    urlTemplate?: string;
+    datafileAccessToken?: string;
+  }
+
   // The options object given to Optimizely.createInstance.
   export interface Config {
     // TODO[OASIS-6649]: Don't use object type
     // eslint-disable-next-line  @typescript-eslint/ban-types
     datafile?: object | string;
-    datafileOptions?: import('./shared_types').DatafileOptions;
+    datafileOptions?: DatafileOptions;
     errorHandler?: ErrorHandler;
     eventDispatcher?: EventDispatcher;
     logger?: LogHandler;
@@ -66,7 +73,7 @@ declare module '@optimizely/optimizely-sdk' {
       eventKey: string,
       userId: string,
       attributes?: import('./shared_types').UserAttributes,
-      eventTags?: import('./shared_types').EventTags
+      eventTags?: EventTags
     ): void;
     getVariation(
       experimentKey: string,
@@ -130,6 +137,34 @@ declare module '@optimizely/optimizely-sdk' {
     close(): Promise<{ success: boolean; reason?: string }>;
   }
 
+  // An event to be submitted to Optimizely, enabling tracking the reach and impact of
+  // tests and feature rollouts.
+  export interface Event {
+    // URL to which to send the HTTP request.
+    url: string;
+    // HTTP method with which to send the event.
+    httpVerb: 'POST';
+    // Value to send in the request body, JSON-serialized.
+    // TODO[OASIS-6649]: Don't use any type
+    // eslint-disable-next-line  @typescript-eslint/no-explicit-any
+    params: any;
+  }
+
+  export interface EventDispatcher {
+    /**
+     * @param event
+     *        Event being submitted for eventual dispatch.
+     * @param callback
+     *        After the event has at least been queued for dispatch, call this function to return
+     *        control back to the Client.
+     */
+    dispatchEvent: (event: Event, callback: () => void) => void;
+  }
+
+  export type EventTags = {
+    [key: string]: string | number | boolean;
+  };
+
   // NotificationCenter-related types
   export interface NotificationCenter {
     addNotificationListener<T extends ListenerPayload>(
@@ -149,14 +184,14 @@ declare module '@optimizely/optimizely-sdk' {
   }
 
   export interface ActivateListenerPayload extends ListenerPayload {
-    experiment: import('./shared_types').Experiment;
-    variation: import('./shared_types').Variation;
+    experiment: import('./core/project_config/entities').Experiment;
+    variation: import('./core/project_config/entities').Variation;
     logEvent: Event;
   }
 
   export interface TrackListenerPayload extends ListenerPayload {
     eventKey: string;
-    eventTags: import('./shared_types').EventTags;
+    eventTags: EventTags;
     logEvent: Event;
   }
 }
